+++ backend/server.py
@@ -1,6 +1,4 @@
 from flask import Flask
-from local_db.orm import User, Ingredient, Recipe, RecipeHasIngredient, RecipeHasWeeklyMenu, RecipeAvailability, \
-    Usergroup, Usertype, WeeklyMenu, Base, sessionmaker
 from local_db.local_db_create import engine
 from sqlalchemy.orm import session
 from local_db.local_db_create import DB_NAME
@@ -12,11 +10,6 @@ app.config['SQLALCHEMY_DATABASE_URI'] = f"mysql+pymysql://{username}:{password}@
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False  # stops warning
 
 
-def loadSession():
-    metadata = Base.metadata
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    return session
 
 
 
@@ -28,9 +21,5 @@ def hello_world():  # put application's code here
 
 if __name__ == '__main__':
 
-    #Test quieries
-    session = loadSession()
-    res = session.query(User).where('username', 'Melvin')
-    print(res[0].email)
 
     app.run()
